:recycle: [tests] Move `commit` to tests.util.git
