+++ tests/functional/test_cookiecutter.py
@@ -5,8 +5,8 @@ import pygit2
 from click.testing import CliRunner
 
 from cutty.entrypoints.cli import main
-from tests.functional.conftest import commit
 from tests.functional.conftest import move_repository_files_to_subdirectory
+from tests.util.git import commit
 
 
 def test_create_help(runner: CliRunner) -> None:
